% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add}
\alias{add}
\alias{add.dendlist}
\alias{add.dendrogram}
\title{Add (/update) features to a dendrogram}
\usage{
add(object, ...)

\method{add}{dendrogram}(object,
   what = c("labels",
            "labels_colors",
            "labels_cex",
            "leaves_pch",
            "leaves_cex",
            "leaves_col",
            "hang_leaves",
            "branches_k_color",
            "branches_col",
            "branches_lwd",
            "branches_lty",
            "clear_branches",
            "clear_leaves"
   ),
   value, ...)

\method{add}{dendlist}(object, ..., which)
}
\arguments{
\item{object}{a tree (\link{dendrogram}, or \link{dendlist})}

\item{what}{a character indicating what is the property of
the tree that should be added/updated. (see the usage and the example section
for the different options)}

\item{value}{an object with the value to add to the tree.
(the type of the value depends on the "what")}

\item{...}{passed to the specific function for more options.}

\item{which}{an integer vector indicating, in the case "object" is
a dendlist, on which of the trees should the modification be performed.
If missing - the change will be performed on all of objects in the dendlist.}
}
\value{
An updated dendrogram (or dendlist), with some change to
the parameters of it
}
\description{
a master function for updating various attributes and
features of dendrogram objects.
}
\details{
This is a wrapper function for many of the main tasks we
might wish to perform on a dendrogram before plotting.
}
\examples{
\dontrun{

set.seed(23235)
ss <- sample(1:150, 10 )

# Getting the dend object
dend <- iris[ss,-5] \%>\% dist \%>\% hclust \%>\% as.dendrogram
dend \%>\% plot

dend \%>\% labels
dend \%>\% add("labels", 1:10) \%>\% labels
dend \%>\% add("labels", 1:10) \%>\% plot
dend \%>\% add("labels_color") \%>\% plot
dend \%>\% add("labels_col", c(1,2)) \%>\% plot # Works also with partial matching :)
dend \%>\% add("labels_cex", c(1, 1.2)) \%>\% plot
dend \%>\% add("leaves_pch", NA) \%>\% plot
dend \%>\% add("leaves_pch", c(1:5)) \%>\% plot
dend \%>\% add("leaves_pch", c(19,19, NA)) \%>\%
   add("leaves_cex", c(1,2)) \%>\% plot
dend \%>\% add("leaves_pch", c(19,19, NA)) \%>\%
   add("leaves_cex", c(1,2)) \%>\%
   add("leaves_col", c(1,1,2,2)) \%>\%
   plot
dend \%>\% add("hang") \%>\% plot

dend \%>\% add("branches_k_col") \%>\% plot
dend \%>\% add("branches_k_col", c(1,2)) \%>\% plot
dend \%>\% add("branches_k_col", c(1,2,3), k=3) \%>\% plot
dend \%>\% add("branches_k_col", k=3) \%>\% plot

dend \%>\% add("branches_col", c(1,2, 1, 2, NA)) \%>\% plot
dend \%>\% add("branches_lwd", c(2,1,2)) \%>\% plot
dend \%>\% add("branches_lty", c(1,2,1)) \%>\% plot

#    clears all of the things added to the leaves
dend \%>\%
   add("labels_color", c(19,19, NA)) \%>\%
   add("leaves_pch", c(19,19, NA))  \%>\%  # plot
   add("clear_leaves") \%>\% # remove all of what was done until this point
   plot
# Different order
dend \%>\%
   add("leaves_pch", c(19,19, NA)) \%>\%
   add("labels_color", c(19,19, NA)) \%>\%
   add("clear_leaves") \%>\% plot


# doing this without chaining (\%>\%) will NOT be fun:
dend \%>\%
   add("labels", 1:10) \%>\%
   add("labels_color") \%>\%
   add("branches_col", c(1,2, 1, 2, NA)) \%>\%
   add("branches_lwd", c(2,1,2)) \%>\%
   add("branches_lty", c(1,2,1)) \%>\%
   add("hang") \%>\%
   plot

# A few dendlist examples:
dendlist(dend,dend) \%>\% add("hang") \%>\% plot
dendlist(dend,dend) \%>\% add("branches_k_col", k=3) \%>\% plot
dendlist(dend,dend) \%>\% add("labels_col", c(1,2)) \%>\% plot

dendlist(dend,dend) \%>\%
   add("hang") \%>\%
   add("labels_col", c(1,2), which = 1) \%>\%
   add("branches_k_col", k=3, which = 2) \%>\%
   add("labels_cex", 1.2) \%>\%
   plot

}
}
\seealso{
\link{labels<-.dendrogram}, \link{labels_colors<-},
\link{hang.dendrogram}, \link{color_branches},
\link{assign_values_to_leaves_nodePar},
\link{assign_values_to_branches_edgePar},
\link{remove_branches_edgePar}, \link{remove_leaves_nodePar},
\link{dendrogram}
}

