% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.phylo.dendrogram}
\alias{as.phylo.dendrogram}
\title{Convert a dendrogram into phylo}
\usage{
as.phylo.dendrogram(x, ...)
}
\arguments{
\item{x}{a dendrogram}

\item{...}{ignored.}
}
\value{
A phylo class object
}
\description{
Based on \link{as.hclust.dendrogram} with \link[ape]{as.phylo.hclust}

In the future I hope a more direct link will be made.
}
\examples{
\dontrun{

library(dendextend)
library(ape)
dend <- 1:50 \%>\% dist \%>\% hclust \%>\% as.dendrogram
dend2 <- as.phylo(dend)
plot(dend2, type = "fan")

}
}
\seealso{
\link{as.dendrogram}, \link{as.hclust}, \link{as.phylo}
}

