% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendextend_options.R
\name{assign_dendextend_options}
\alias{assign_dendextend_options}
\title{Populates dendextend functions into dendextend_options}
\usage{
assign_dendextend_options()
}
\description{
Populates dendextend functions into dendextend_options
}
