\name{bakers_gamma_for_2_k_matrix}
\alias{bakers_gamma_for_2_k_matrix}
\title{Bakers Gamma for two k matrices}
\usage{
bakers_gamma_for_2_k_matrix(k_matrix_tree1, k_matrix_tree2, to_plot = FALSE)
}
\arguments{
  \item{k_matrix_tree1}{a matrix of k cluster groupings
  from a dendrogram}

  \item{k_matrix_tree1}{a matrix of k cluster groupings
  from a dendrogram}

  \item{to_plot}{logical (FALSE). Should a scaterplot be
  plotted, showing the correlation between the lowest
  shared branch between two items in the two compared
  trees.}

  \item{...}{not used}
}
\value{
Baker's Gamma coefficient.
}
\description{
Bakers Gamma for two k matrices
}
\seealso{
\link{cor_bakers_gamma}
}

