% Generated by roxygen2 (4.0.1): do not edit by hand
\name{click_rotate}
\alias{click_rotate}
\alias{click_rotate.default}
\alias{click_rotate.dendrogram}
\title{Interactively rotate a tree object}
\usage{
click_rotate(x, ...)

\method{click_rotate}{dendrogram}(x, plot = TRUE, plot_after = plot, ...)
}
\arguments{
\item{x}{a tree object (either a \code{dendrogram} or \code{hclust})}

\item{plot}{(logical) should the dendrogram first be plotted.}

\item{plot_after}{(logical) should the dendrogram be plotted after
the rotation?}

\item{...}{parameters passed to the plot}
}
\value{
A rotated tree object
}
\description{
Lets te user click a plot of dendrogram
and rotates the tree based on the location of the click.
}
\examples{
# create the dend:
dend <- USArrests \%>\% dist \%>\% hclust("ave") \%>\%
as.dendrogram \%>\% color_labels

\dontrun{
# play with the rotation once
dend <- click_rotate(dend)
# keep playing with the rotation:
while(TRUE) dend <- click_rotate(dend)
}
}
\author{
Andrej-Nikolai Spiess, Tal Galili
}
\seealso{
\code{\link{rotate.dendrogram}}
}

