% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_colors.R
\name{color_unique_labels}
\alias{color_unique_labels}
\title{Color unique labels in a dendrogram}
\usage{
color_unique_labels(dend, ...)
}
\arguments{
\item{dend}{a dend object}

\item{...}{NOT USED}
}
\value{
A dendrogram after the colors of its labels have been updated (a different color for each unique label).
}
\description{
Color unique labels in a dendrogram
}
\examples{

x <- c(2011, 2011, 2012, 2012, 2015, 2015, 2015)
names(x) <- x
dend <- as.dendrogram(hclust(dist(x)))

par(mfrow = c(1, 2))
plot(dend)
dend2 <- color_unique_labels(dend)
plot(dend2)
}
