% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dend_diff}
\alias{dend_diff}
\title{Plots two trees side by side, highlighting edges unique to each tree in red.}
\source{
A \link{dendrogram} implementation for \link[distory]{phylo.diff} from the {distory} package
}
\usage{
dend_diff(x, y, horiz = TRUE, ...)
}
\arguments{
\item{x}{a dendrogram to find unique edges in}

\item{y}{a dendrogram to compare with}

\item{horiz}{logical (TRUE) indicating if the dendrogram should be drawn horizontally or not.}

\item{...}{passed to \link{plot.dendrogram}}
}
\value{
Invisible NULL
}
\description{
Plots two trees side by side, highlighting edges unique to each tree in red.
}
\examples{
x <- 1:5 \%>\% dist \%>\% hclust \%>\% as.dendrogram
y <- set(x, "labels", 5:1)

dend_diff(x, y)

dend1 <- 1:10 \%>\% dist \%>\% hclust \%>\% as.dendrogram
dend2 <- dend1 \%>\% set("labels", c(1,3,2,4, 5:10) )
dend_diff(dend1, dend2)
}
\seealso{
\link{distinct_edges}, \link{highlight_distinct_edges},
\link{edgeset_dist}, \link{dist.dendlist}, \link{tanglegram}
\link{assign_values_to_branches_edgePar},
\link[distory]{distinct.edges},
}

