\docType{package}
\name{dendextend-package}
\alias{dendextend}
\alias{dendextend-package}
\title{Functions for extending dendrogram objsects}
\description{
  Functions for extending dendrogram objsects
}
\seealso{
  \code{\link{dendrogram}, \link{hclust}} in
  \code{\link{stats}} package.
}
\keyword{dendrogram}
\keyword{package,}

