% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_imports.R
\name{dendextend_labels.dendrogram}
\alias{dendextend_labels.dendrogram}
\title{Labels of the Leaves in a Dendrogram}
\usage{
dendextend_labels.dendrogram(object, ...)
}
\arguments{
\item{object}{a \link{dendrogram}.}

\item{...}{not used.}
}
\value{
A vector with length equal to the number of leaves in the dendrogram is returned. From r <- order.dendrogram(), each element is the index into the original data (from which the dendrogram was computed).
}
\description{
This is the same function as \link[stats]{labels.dendrogram}. It is used inside
 \link{dendextend_options}, and thus allows the function to be updated when loading
 the dendextendRcpp package.
}
\examples{

# stats:::labels.dendrogram

}
\seealso{
\link[stats]{labels.dendrogram}
}
