% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_access.R
\name{get_childrens_heights}
\alias{get_childrens_heights}
\title{Get height attributes from a dendrogram's children}
\usage{
get_childrens_heights(dend, ...)
}
\arguments{
\item{dend}{a dendrogram.}

\item{...}{not used.}
}
\value{
a vector of the heights of a dendrogram's current node's
(first level) children.
}
\description{
Get height attributes from a dendrogram's children nodes
}
\examples{

hc <- hclust(dist(USArrests[1:4, ]), "ave")
dend <- as.dendrogram(hc)
get_childrens_heights(dend)
}
\seealso{
\link{get_branches_heights}
}
