% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendlist.R
\name{is.dendlist}
\alias{is.dendlist}
\title{Checks if an object is of class dendlist}
\usage{
is.dendlist(x)
}
\arguments{
\item{x}{whatever object to check}
}
\value{
logical
}
\description{
Checks if an object is of class dendlist
}
\examples{
# TRUE:
is.dendlist(dendlist())
# FALSE
is.dendlist(1)
}

