% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutree.dendrogram.R
\name{is.dendrogram}
\alias{is.dendrogram}
\title{Is the object of class dendrogram}
\usage{
is.dendrogram(x)
}
\arguments{
\item{x}{an object.}
}
\value{
logical - is the object of class dendrogram.
}
\description{
Is the object of class dendrogram
}

