% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.functions.R
\name{is.dist}
\alias{is.dist}
\title{Checks if an object is of class dist}
\usage{
is.dist(x)
}
\arguments{
\item{x}{whatever object to check}
}
\value{
logical
}
\description{
Checks if an object is of class dist (returns TRUE if it is).
}
\examples{
# TRUE:
is.dist(dist(mtcars))
# FALSE
is.dist(mtcars)
}

