% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cutree.dendrogram.R
\name{is.phylo}
\alias{is.phylo}
\title{Is the object of class phylo}
\usage{
is.phylo(x)
}
\arguments{
\item{x}{an object.}
}
\value{
logical - is the object of class phylo.
}
\description{
Is the object of class phylo
}

