% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendlist.R
\name{is_null_list}
\alias{is_null_list}
\title{Checks if the value is and empty list()}
\usage{
is_null_list(x)
}
\arguments{
\item{x}{whatever object to check}
}
\value{
logical
}
\description{
Checks if the value is and empty list(). Can be useful.
}
\examples{
# I can run this only if I'd make is_null_list exported
\dontrun{
# TRUE:
is_null_list(list())
# FALSE
is_null_list(list(1))
is_null_list(1)

x <- list(1, list(), 123)
ss_list <- sapply(x, is_null_list)
x <- x[!ss_list]
x

x <- list(1, list(), 123)
ss_list <- sapply(x, is_null_list)
x <- list(list())
x
}

\dontrun{
# error
is_null_list()
}
}
