% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.functions.R
\name{is_some_class}
\alias{is_some_class}
\alias{is.hclust}
\alias{is.dendrogram}
\alias{is.phylo}
\alias{is.dendlist}
\alias{is.dist}
\title{Is the object of some class}
\usage{
is.hclust(x)

is.dendrogram(x)

is.phylo(x)

is.dendlist(x)

is.dist(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Returns TRUE if some class (based on the name of the function).
}
\description{
Returns TRUE if some class (based on the name of the function).
}
\examples{
# TRUE:
is.dendlist(dendlist())
# FALSE
is.dendlist(1)
# TRUE:
is.dist(dist(mtcars))
# FALSE
is.dist(mtcars)
}
