% Generated by roxygen2 (4.0.1): do not edit by hand
\name{leaf_Colors}
\alias{leaf_Colors}
\alias{leaf_colors}
\title{Return the leaf Colors of a dendrogram}
\usage{
leaf_Colors(d, col_to_return = c("edge", "node", "label"))
}
\arguments{
\item{d}{the dendrogram}

\item{col_to_return}{Character scalar - kind of Color attribute to return}
}
\value{
named character vector of Colors, NA_character_ where missing
}
\description{
Return the leaf Colors of a dendrogram
}
\details{
The returned Colors will be in dendrogram order.
}
\examples{
dend <- as.dendrogram(hclust(dist(USArrests), "ave"))
d5=color_branches(dend,5)
leaf_Colors(d5)
}
\author{
jefferis
}
\seealso{
\code{\link[dendroextras]{slice},\link{color_branches}}
}

