% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{order.hclust}
\alias{order.hclust}
\title{Ordering of the Leaves in a hclust Dendrogram}
\usage{
order.hclust(x, ...)
}
\arguments{
\item{x}{ab hclust object a distance matrix.}

\item{...}{Ignored.}
}
\value{
A vector with length equal to the number of
leaves in the hclust dendrogram is returned.
From r <- order.hclust(), each element is
the index into the original data
(from which the hclust was computed).
}
\description{
Ordering of the Leaves in a hclust Dendrogram. Like \link{order.dendrogram}.
}
\examples{

set.seed(23235)
ss <- sample(1:150, 10)
hc <- iris[ss, -5] \%>\%
  dist() \%>\%
  hclust()
# dend <- hc \%>\% as.dendrogram
order.hclust(hc)
}
\seealso{
\link{order.dendrogram}
}
