% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/magrittr.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe an object forward into a function call/expression.}
\source{
This is just the imported function
from the magrittr package. The documentation you should
read for the \%>\% function can be found here: \link[magrittr]{pipe}

Adding the function in the package while importing it
is a trick taken from the {dplyr} package
(in the file chain.r)

The package could have also been placed in "Depends",
But I wanted to keep it robust if at some point some other
package will try to take over the symbol (not likely, but still).
}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The value to be piped}

\item{rhs}{A function or expression}
}
\description{
Pipe an object forward into a function call/expression.
}
\seealso{
\link[magrittr]{pipe}
}

