% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_common_subtrees.dendlist}
\alias{prune_common_subtrees.dendlist}
\title{Prune trees to their common subtrees}
\usage{
prune_common_subtrees.dendlist(dend, ...)
}
\arguments{
\item{dend}{a \link{dendlist} of length two}

\item{...}{ignored}
}
\value{
A dendlist after prunning the labels to only include
those that are part of common subtrees in both dendrograms.
}
\description{
Prune trees to their common subtrees
}
\examples{

# NULL
}
\seealso{
\link{common_subtrees_clusters}
}
