% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvclust.R
\name{pvclust_edges}
\alias{pvclust_edges}
\title{Get Pvclust Edges Information}
\usage{
pvclust_edges(pvclust_obj)
}
\arguments{
\item{pvclust_obj}{pvclust object}
}
\value{
data.frame with leaves on column 1 and 2, followed by the rest of the information from edge
}
\description{
Get pvclust edges information such as au and bp and return dataframe with proper sample labels. This function is useful when there is a lot of samples involved.
}
\examples{
pvclust.edges(pvclust_obj)

}
\references{
hclust object descriptions \url{https://stat.ethz.ch/R-manual/R-patched/library/stats/html/hclust.html}
}
