% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_nodes_nodePar}
\alias{remove_nodes_nodePar}
\title{Remove all nodePar values from a dendrogram's nodes}
\usage{
remove_nodes_nodePar(object, ...)
}
\arguments{
\item{object}{a dendrogram object}

\item{...}{not used}
}
\value{
A dendrogram, after removing the nodePar attribute in all of its nodes,
}
\description{
Go through the dendrogram nodes and remove its nodePar
}
\examples{
\dontrun{

dend <- USArrests[1:5,] \%>\% dist \%>\% hclust \%>\% as.dendrogram
dend <- color_branches(dend, 3)
par(mfrow = c(1,2))
plot(dend)
plot(remove_branches_edgePar(dend))

}
}
\seealso{
\link{get_root_branches_attr}, \link{assign_values_to_branches_edgePar}
}

