% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cophenetic.R
\name{sort_dist_mat}
\alias{sort_dist_mat}
\title{Sorts a distance matrix by rows and columns names}
\usage{
sort_dist_mat(dist_mat, by_rows = TRUE, by_cols = TRUE, ...)
}
\arguments{
\item{dist_mat}{a distance matrix.}

\item{by_rows}{logical (TRUE). Sort the distance matrix by rows?}

\item{by_cols}{logical (TRUE). Sort the distance matrix by columns?}

\item{...}{Ignored.}
}
\value{
A distance matrix (after sorting)
}
\description{
Sorts a distance matrix by the names of the rows and columns.
}
\seealso{
\link{dist}, \link{cor_cophenetic}
}
