% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdendro.R
\name{theme_dendro}
\alias{theme_dendro}
\title{Creates completely blank theme in ggplot}
\source{
This function is from Andrie de Vries's ggdendro package.

The motivation for this fork is the need to add more graphical parameters to the plotted tree.
This required a strong mixter of functions from ggdendro and dendextend (to the point that
it seemed better to just fork the code into its current form)
}
\usage{
theme_dendro()
}
\description{
Sets most of the ggplot options to blank, by returning blank theme elements for the panel grid, panel background, axis title, axis text, axis line and axis ticks.
}
\seealso{
\link{ggdend}
}
\author{
Andrie de Vries
}
