\docType{package}
\name{dendextend-package}
\alias{dendextend}
\alias{dendextend-package}
\title{Functions for extending dendrogram objects}
\description{
Functions for extending dendrogram objects
}
\seealso{
\code{\link{dendrogram}, \link{hclust}} in
\code{\link{stats}} package.
}
\keyword{dendrogram}
\keyword{package,}

