% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendextend-package.R
\docType{package}
\name{dendextend-package}
\alias{dendextend-package}
\alias{dendextend}
\title{Functions for extending dendrogram objects}
\description{
Functions for extending dendrogram objects
}
\seealso{
\code{\link{dendrogram}, \link{hclust}} in \code{\link{stats}} package.
}
\keyword{clustering}
\keyword{dendrogram}
\keyword{package}
\keyword{visualization}
