% Generated by roxygen2 (4.0.2): do not edit by hand
\name{highlight_distinct_edges}
\alias{highlight_distinct_edges}
\title{Highlight distint edges in a tree (compared to another one)}
\usage{
highlight_distinct_edges(x, y, value = 2, edgePar = c("col", "lty", "lwd"),
  ...)
}
\arguments{
\item{x}{a dendrogram to find unique edges in}

\item{y}{a dendrogram to compare with}

\item{value}{a new value scalar for the edgePar attribute.}

\item{edgePar}{a character indicating the value inside edgePar to adjust.
Can be either "col", "lty", or "lwd".}

\item{...}{Ignored.}
}
\value{
A dendrogram with modified edges - the distinct ones are changed (color, line width, or line type)
}
\description{
Highlight distint edges in a tree (compared to another one) by changing
the branches' color, line width, or line type.

This function enables this feature in \link{dend_diff} and \link{tanglegram}
}
\examples{
x <- 1:5 \%>\% dist \%>\% hclust \%>\% as.dendrogram
y <- set(x, "labels", 5:1)
distinct_edges(x, y)
distinct_edges(y, x)

par(mfrow = c(1,2))
plot(highlight_distinct_edges(x, y))
plot(y)

# tanglegram(highlight_distinct_edges(x, y),y)
# dend_diff(x, y)
}
\seealso{
\link{distinct_edges}, \link{highlight_distinct_edges},
\link{dist.dendlist}, \link{tanglegram}
\link{assign_values_to_branches_edgePar},
\link[distory]{distinct.edges},
}

