% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.functions.R
\name{is.hclust}
\alias{is.hclust}
\title{Is the object of class hclust}
\usage{
is.hclust(x)
}
\arguments{
\item{x}{an object.}
}
\value{
logical - is the object of class hclust.
}
\description{
Is the object of class hclust.
}

