% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_access.R
\name{remove_leaves_nodePar}
\alias{remove_leaves_nodePar}
\title{Remove all nodePar values from a dendrogram's leaves}
\usage{
remove_leaves_nodePar(dend, ...)
}
\arguments{
\item{dend}{a dendrogram object}

\item{...}{not used}
}
\value{
A dendrogram, after removing the nodePar attribute in all of its leaves,
}
\description{
Go through the dendrogram leaves and remove its nodePar.
}
\examples{

\dontrun{

dend <- USArrests[1:5, ] \%>\%
  dist() \%>\%
  hclust() \%>\%
  as.dendrogram()

dend <- color_labels(dend, 3)
par(mfrow = c(1, 2))
plot(dend)
plot(remove_leaves_nodePar(dend))


get_leaves_attr(dend, "nodePar")
get_leaves_attr(remove_leaves_nodePar(dend), "nodePar")
}

}
\seealso{
\link{get_leaves_attr}, \link{assign_values_to_leaves_nodePar}
}
